chore: move test utils into src directory
This should create a simpler project structure, with everything living under the `src` directory.
