+++ jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  testMatch: ["**/*.test.ts"],
+  testMatch: ["<rootDir>/src/**/*.test.ts"],
   transform: {
     "^.+\\.tsx?$": "ts-jest",
   },
