+++ src/constructs/core/parameters/base.test.ts
@@ -1,7 +1,7 @@
 import "@aws-cdk/assert/jest";
 import { SynthUtils } from "@aws-cdk/assert/lib/synth-utils";
-import type { SynthedStack } from "../../../../test/utils";
-import { simpleGuStackForTesting } from "../../../../test/utils";
+import type { SynthedStack } from "../../../utils/test";
+import { simpleGuStackForTesting } from "../../../utils/test";
 import { GuArnParameter, GuParameter, GuStringParameter } from "./base";
 
 describe("The GuParameter class", () => {
