+++ src/constructs/core/parameters/ec2.test.ts
@@ -1,7 +1,7 @@
 import "@aws-cdk/assert/jest";
 import { SynthUtils } from "@aws-cdk/assert/lib/synth-utils";
-import type { SynthedStack } from "../../../../test/utils";
-import { simpleGuStackForTesting } from "../../../../test/utils";
+import { simpleGuStackForTesting } from "../../../utils/test";
+import type { SynthedStack } from "../../../utils/test";
 import { GuInstanceTypeParameter } from "./ec2";
 
 describe("The GuInstanceTypeParameter class", () => {
