+++ src/constructs/iam/policies/assume-role.test.ts
@@ -1,5 +1,5 @@
 import "@aws-cdk/assert/jest";
-import { attachPolicyToTestRole, simpleGuStackForTesting } from "../../../../test/utils";
+import { attachPolicyToTestRole, simpleGuStackForTesting } from "../../../utils/test";
 import { GuAssumeRolePolicy } from "./assume-role";
 
 describe("The GuAssumeRolePolicy class", () => {
