+++ src/constructs/iam/policies/describe-ec2.test.ts
@@ -1,5 +1,5 @@
 import "@aws-cdk/assert/jest";
-import { attachPolicyToTestRole, simpleGuStackForTesting } from "../../../../test/utils";
+import { attachPolicyToTestRole, simpleGuStackForTesting } from "../../../utils/test";
 import { GuDescribeEC2Policy } from "./describe-ec2";
 
 describe("DescribeEC2Policy", () => {
