+++ src/constructs/iam/policies/s3-get-object.test.ts
@@ -1,7 +1,7 @@
 import "@aws-cdk/assert/jest";
 import { SynthUtils } from "@aws-cdk/assert/lib/synth-utils";
-import type { SynthedStack } from "../../../../test/utils";
-import { attachPolicyToTestRole, simpleGuStackForTesting } from "../../../../test/utils";
+import { attachPolicyToTestRole, simpleGuStackForTesting } from "../../../utils/test";
+import type { SynthedStack } from "../../../utils/test";
 import { GuDistributionBucketParameter } from "../../core";
 import { GuGetDistributablePolicy, GuGetS3ObjectsPolicy } from "./s3-get-object";
