+++ src/constructs/loadbalancing/alb.test.ts
@@ -3,9 +3,9 @@ import { SynthUtils } from "@aws-cdk/assert/lib/synth-utils";
 import { Vpc } from "@aws-cdk/aws-ec2";
 import { ApplicationProtocol, ListenerAction } from "@aws-cdk/aws-elasticloadbalancingv2";
 import { Stack } from "@aws-cdk/core";
-import { simpleGuStackForTesting } from "../../../test/utils";
-import type { SynthedStack } from "../../../test/utils";
 import { RegexPattern } from "../../constants";
+import { simpleGuStackForTesting } from "../../utils/test";
+import type { SynthedStack } from "../../utils/test";
 import type { AppIdentity } from "../core/identity";
 import {
   GuApplicationListener,
