+++ src/constructs/loadbalancing/elb.test.ts
@@ -2,8 +2,8 @@ import "@aws-cdk/assert/jest";
 import { SynthUtils } from "@aws-cdk/assert/lib/synth-utils";
 import { Vpc } from "@aws-cdk/aws-ec2";
 import { Stack } from "@aws-cdk/core";
-import { simpleGuStackForTesting } from "../../../test/utils";
-import type { SynthedStack } from "../../../test/utils";
+import { simpleGuStackForTesting } from "../../utils/test";
+import type { SynthedStack } from "../../utils/test";
 import { GuClassicLoadBalancer, GuHttpsClassicLoadBalancer } from "./elb";
 
 describe("The GuClassicLoadBalancer class", () => {
