+++ src/utils/test/simple-gu-stack.ts
@@ -1,6 +1,6 @@
 import { App } from "@aws-cdk/core";
-import type { GuStackProps } from "../../src/constructs/core";
-import { GuStack } from "../../src/constructs/core";
+import type { GuStackProps } from "../../constructs/core";
+import { GuStack } from "../../constructs/core";
 
 export const simpleGuStackForTesting: (props?: Partial<GuStackProps>) => GuStack = (props?: Partial<GuStackProps>) =>
   new GuStack(new App(), "Test", { stack: props?.stack ?? "test-stack", ...props });
