+++ tsconfig.eslint.json
@@ -2,7 +2,6 @@
   "extends": "./tsconfig.json",
   "include": [
     "src/**/*",
-    "test/**/*",
     "eslint/rules/valid-constructors.test.ts"
   ],
   "exclude": [
